Moved AsyncSemanticScholarTitleAbstractEmbeddingVectorProvider to providers
